Stack Overflow solved by changing int to int64_t in UPC because the codes were too large for it to handle. Changed Node.h destructor to change left and right pointers to nullptr

+++ UPCScanner/UPCScanner/UPC.h
@@ -19,7 +19,7 @@ class UPC
 			key = _key;
 		}
 
-		UPC(int& _key, std::string& _value)
+		UPC(int64_t& _key, std::string& _value)
 		{
 			key = _key;
 			value = _value;
@@ -43,7 +43,7 @@ class UPC
 		}
 
 	private:
-		int key;
+		int64_t key;
 		std::string value;
 };
